❗️HomeViewの段階で画面遷移するとデータが消える
