+++ Ibum/PhotoTrimmingView.swift
@@ -9,12 +9,14 @@ import SwiftData
 
 struct PhotoTrimmingView: View {
     
+    @Environment(\.modelContext) private var context
+    @Query var quests: [Quest]
     
     @State private var scale: CGFloat = 1.01
     @State var position: CGPoint = CGPoint(x: UIScreen.main.bounds.width / 2, y: UIScreen.main.bounds.height / 2)
     
     @State var image:UIImage
-    @State var questTitle:String = "笑顔でピース"
+    @State var questTitle:String = ""
     
    
 
@@ -96,6 +98,7 @@ struct PhotoTrimmingView: View {
         }
         .onAppear{
             do{
+                print(quests.first)
             }catch{
                 print(error)
             }
@@ -110,21 +113,20 @@ struct PhotoTrimmingView: View {
 //        }
 //    }
 //    
-//    @Environment(\.modelContext) private var context
-//    @Query var quests: [Quest]
+
 //    
     func add() {
-//        let uuid =  UUID()
-//        do{
+        let uuid =  UUID()
+        do{
 //            let descriptor = FetchDescriptor<Quest>(predicate: #Predicate<Quest>{$0.title == "笑顔でピース"})
 //            let currentQuest = try context.fetch(descriptor).first
-//            let data = Photo(saveDate: Date(), photoData: image.jpegData(compressionQuality: 1)!, scale: scale, center: position, registerSns: [], best: true,questTitle: "笑顔でピース",id: uuid)
-//            context.insert(data)
-//            try context.save()
-//            print(currentQuest)
-//        }catch{
-//            print(error)
-//        }
+            let data = Photo(saveDate: Date(), photoData: image.jpegData(compressionQuality: 1)!, scale: scale, center: position, registerSns: [], best: true,questTitle: questTitle,id: uuid)
+            context.insert(data)
+            try context.save()
+            print(quests)
+        }catch{
+            print(error)
+        }
 //
